# design/API.yml
openapi: '3.0.2'
info:
  title: Cheese Community API
  version: '1.0'
servers:
  - url: http://localhost:7777/api/{version}
    description: Local server
    variables:
      version:
        default: v1
components:
  parameters:
...
    groupId:
      name: groupId
      in: path
      description: 圈子 ID
      required: true
      schema:
        type: integer
        format: int64
...
    page_start:
      name: page_start
      in: query
      description: 该页第一个 item 的 ID，留空即为从头开始
      schema:
        type: integer
        format: int64
    page_size:
      name: page_size
      in: query
      description: 每页 item 数量
      schema:
        type: integer
        format: int64
        default: 20
  securitySchemes:
    bearerAuth:
      description: The access token is used in this way.
      type: http
      scheme: bearer
      bearerFormat: JWT
    bearerAuth2:
      type: apiKey
      description: The refresh token is used in this way.
      name: REFRESH_TOKEN
      in: cookie
  schemas:
    CommonResponse:
      type: object
      properties:
        code:
          type: integer
          format: int32
          default: 0
          description: 错误码，0表示成功，其他表示失败
        message:
          type: string
          default: ''
          description: 错误信息，成功时为空字符串
      required:
        - code
        - message
    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: 用户 ID
          minimum: 1
        username:
          type: string
          description: 用户名
          example: 'cheese'
        nickname:
          type: string
          description: 昵称
          example: '芝士'
        avatar:
          type: string
          description: 头像 URL
        intro:
          type: string
          description: 个人简介
          default: 'This user has not set an introduction yet.'
        follow_count:
          type: integer
          format: int64
          description: 关注用户的数量
          minimum: 0
          example: 114
        fans_count:
          type: integer
          format: int64
          description: 粉丝数量
          minimum: 0
          example: 514
        question_count:
          type: integer
          format: int64
          description: 提问数量
          minimum: 0
          example: 1919
        answer_count:
          type: integer
          format: int64
          description: 回答数量
          minimum: 0
          example: 810
        is_follow:
          type: boolean
          description: 当前登录用户是否关注该用户
      required:
        - id
        - username
        - nickname
        - avatar
        - intro
        - follow_count
        - fans_count
        - question_count
        - answer_count
    Page:
      type: object
      description: 分页信息
      properties:
        page_start:
          type: integer
          format: int64
          description: 该页第一个 item 的 ID
        page_size:
          type: integer
          format: int64
          description: 每页 item 数量
        has_prev:
          type: boolean
          description: 是否有上一页
        prev_start:
          type: integer
          format: int64
          description: 上一页第一个 item 的 ID
        has_more:
          type: boolean
          description: 是否有下一页
        next_start:
          type: integer
          format: int64
          description: 下一页第一个 item 的 ID
    Topic:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
      required:
        - id
        - name
    Question:
      type: object
      description: 问题详情
      properties:
        id:
          type: integer
          format: int64
          description: 问题 ID
        title:
          type: string
          description: 问题标题
        content:
          type: string
          description: 问题详情
        author:
          $ref: '#/components/schemas/User'
        type:
          type: integer
          description: 问题类型
        topics:
          type: array
          items:
            $ref: '#/components/schemas/Topic'
        created_at:
          type: integer
          format: int64
          description: 创建时间（时间戳）
        updated_at:
          type: integer
          format: int64
          description: 最后修改时间（时间戳）
        is_follow:
          type: boolean
          description: 是否关注问题
        is_like:
          type: boolean
          description: 是否点赞问题
        answer_count:
          type: integer
          description: 回答数量
        comment_count:
          type: integer
          description: 评论数量
        follow_count:
          type: integer
          description: 关注数量
        like_count:
          type: integer
          description: 点赞数量
        view_count:
          type: integer
          description: 浏览数量
        is_group:
          type: boolean
          description: 是否为小组问题
        group:
          $ref: '#/components/schemas/Group'
...
    Group:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        intro:
          type: string
          default: ''
        owner:
          $ref: '#/components/schemas/User'
        created_at:
          type: integer
          format: int64
        updated_at:
          type: integer
          format: int64
        member_count:
          type: integer
          format: int64
        question_count:
          type: integer
          format: int64
        answer_count:
          type: integer
          format: int64
        is_member:
          type: boolean
        is_owner:
          type: boolean
        is_public:
          type: boolean
      required:
        - id
        - name
        - intro
...
paths:
...
  /groups:
    get:
      summary: 获取圈子列表
      operationId: getGroups
      parameters:
        - name: q
          in: query
          description: 搜索关键词，留空即为获取全部圈子
          schema:
            type: string
        - $ref: '#/components/parameters/page_start'
        - $ref: '#/components/parameters/page_size'
        - name: sort
          in: query
          description: 排序方式
          schema:
            type: string
            enum:
              - recommend
              - hot
              - new
            default: recommend
        - name: joined
          in: query
          description: 是否为我参与的小组
          schema:
            type: boolean
        - name: managed
          in: query
          description: 是否为我管理的小组
          schema:
            type: boolean
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                allOf:
                  - $ref: '#/components/schemas/CommonResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          groups:
                            type: array
                            items:
                              $ref: '#/components/schemas/Group'
                          page:
                            $ref: '#/components/schemas/Page'
                    required:
                      - data
    post:
      summary: 创建圈子
      operationId: createGroup
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                intro:
                  type: string
                avatar:
                  type: string
                public:
                  type: boolean
              required:
                - name
                - intro
                - avatar
                - public
      responses:
        '201':
          description: Group created successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/CommonResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          id:
                            type: integer
                            format: int64
                    required:
                      - data
  /groups/{groupId}:
    get:
      summary: 获取圈子详情
      operationId: getGroupDetail
      parameters:
        - $ref: '#/components/parameters/groupId'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/CommonResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          group:
                            $ref: '#/components/schemas/Group'
                    required:
                      - data
    put:
      summary: 更新圈子
      operationId: updateGroup
      parameters:
        - $ref: '#/components/parameters/groupId'
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                intro:
                  type: string
                avatar:
                  type: string
                cover:
                  type: string
              required:
                - name
                - intro
                - avatar
                - cover
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommonResponse'
    delete:
      summary: 删除圈子
      operationId: deleteGroup
      parameters:
        - $ref: '#/components/parameters/groupId'
      security:
        - bearerAuth: []
      responses:
        '204':
          description: No Content
  /groups/{groupId}/members:
    get:
      summary: 获取圈子成员列表
      operationId: getGroupMembers
      parameters:
        - $ref: '#/components/parameters/groupId'
        - $ref: '#/components/parameters/page_start'
        - $ref: '#/components/parameters/page_size'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                allOf:
                  - $ref: '#/components/schemas/CommonResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          members:
                            type: array
                            items:
                              $ref: '#/components/schemas/User'
                          page:
                            $ref: '#/components/schemas/Page'
                    required:
                      - data
    post:
      summary: 加入圈子
      operationId: joinGroup
      parameters:
        - $ref: '#/components/parameters/groupId'
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                intro:
                  type: string
              required:
                - intro
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/CommonResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          member_count:
                            type: integer
                            description: 新的成员数量
                          is_member:
                            type: boolean
                            description: 是否已加入圈子
                          is_waiting:
                            type: boolean
                            description: 是否正在等待审核
                    required:
                      - data
    delete:
      summary: 退出圈子
      operationId: quitGroup
      parameters:
        - $ref: '#/components/parameters/groupId'
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                allOf:
                  - $ref: '#/components/schemas/CommonResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          member_count:
                            type: integer
                            description: 新的成员数量
                    required:
                      - data
  /groups/{groupId}/questions:
    get:
      summary: 获取圈子问题列表
      operationId: getGroupQuestions
      parameters:
        - $ref: '#/components/parameters/groupId'
        - $ref: '#/components/parameters/page_start'
        - $ref: '#/components/parameters/page_size'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                allOf:
                  - $ref: '#/components/schemas/CommonResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          questions:
                            type: array
                            items:
                              $ref: '#/components/schemas/Question'
                          page:
                            $ref: '#/components/schemas/Page'
                    required:
                      - data
...