# src/topics/topics.legacy.entity.ts
/*
 *  DO NOT MODIFY THIS FILE!
 *
 *  TypeORM is deprecated in this project, and we use Prisma instead.
 *  This file is only used by legacy code.
 *
 *  A .legacy.prisma file with same schema is generated by Prisma.
 *  If you modify this file, the schema in .legacy.prisma will be inconsistent
 *  with this file. This may cause unexpected errors.
 *
 *  Also, because TypeORM will alter the database schema automatically if
 *  'syncronize' is set to true, modifying this file is extremely dangerous.
 *
 *  2024-02-19 by Nictheboy <nictheboy@outlook.com>
 *
 */

/*
 *  Description: This file defines the entities used in topics service.
 *               It defines the SQL tables stored in the database.
 *
 *  Author(s):
 *      Nictheboy Li    <nictheboy@outlook.com>
 *
 */

import {
    Column,
    CreateDateColumn,
    DeleteDateColumn,
    Entity,
    Index,
    ManyToOne,
    PrimaryGeneratedColumn,
} from 'typeorm';
import { isMySQL } from '../common/helper/db.helper';
import { User } from '../users/users.legacy.entity';

@Entity()
export class Topic {
  @PrimaryGeneratedColumn()
  id: number;

  @Column()
  @Index('idx_topic_name_ft', { fulltext: true, parser: 'ngram' }) // For fulltext search.
  @Index('idx_topic_name_unique', { unique: true }) // For uniqueness.
  name: string;

  @ManyToOne(() => User)
  @Index()
  createdBy: User;

  // This property does not generate a new column, because the column `createdById` is
  // generated automatically according to the @ManyToOne decorator by TypeORM engine.
  //
  // This property is used for accessing the user id without joining the user table.
  @Column()
  createdById: number;

  @CreateDateColumn()
  createdAt: Date;

  @DeleteDateColumn()
  deletedAt?: Date;
}

@Entity()
// The search history is a precious data source,
// so we should record it even if it is not used for now.
export class TopicSearchLog {
  @PrimaryGeneratedColumn()
  id: number;

  @Column()
  // In the future, we may want to use the search history to recommend topics to users.
  // So we use fulltext index here, although it is not necessary for now.
  @Index({ fulltext: true, parser: 'ngram' })
  keywords: string;

  @Column('int', { nullable: true })
  // A paging argument.
  firstTopicId?: number | null;

  @Column()
  // A paging argument.
  pageSize: number;

  @Column()
  // The result is represented as a string of topic ids, separated by comma.
  // For example, if the result is [1, 2, 3], then the result string is "1,2,3".
  result: string;

  @Column({ type: isMySQL() ? 'double' : 'float' })
  // The search duration in seconds.
  duration: number;

  @ManyToOne(() => User)
  @Index()
  searcher: User;

  // This property does not generate a new column, because the column `searcherId` is
  // generated automatically according to the @ManyToOne decorator by TypeORM engine.
  //
  // This property is used for accessing the user id without joining the user table.
  //
  // undefined if the searcher is not logged in.
  @Column('int', { nullable: true })
  searcherId?: number | null;

  @Column()
  ip: string;

  @Column()
  userAgent: string = '';

  @CreateDateColumn()
  createdAt: Date;
}