# src/users/DTO/user.dto.ts
export class UserDto {
  id: number;
  username: string;
  nickname: string;
  avatar: string;
  intro: string;
  follow_count: number;
  fans_count: number;
  question_count: number;
  answer_count: number;
  is_follow: boolean;
}